+++ 1.20.1/Forge/src/main/java/fuzs/thinair/ThinAirForge.java
@@ -8,7 +8,6 @@ import fuzs.puzzleslib.api.event.v1.core.EventResult;
 import fuzs.puzzleslib.api.event.v1.core.ForgeEventInvokerRegistry;
 import fuzs.puzzleslib.api.event.v1.data.DefaultedInt;
 import fuzs.puzzleslib.api.event.v1.entity.living.LivingEvents;
-import fuzs.thinair.capability.AirBubblePositionsCapability;
 import fuzs.thinair.config.CommonConfig;
 import fuzs.thinair.data.*;
 import fuzs.thinair.init.ForgeModRegistry;
@@ -20,11 +19,14 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraftforge.common.capabilities.CapabilityToken;
 import net.minecraftforge.event.entity.living.LivingBreatheEvent;
 import net.minecraftforge.event.entity.living.LivingDrownEvent;
+import net.minecraftforge.eventbus.api.IEventBus;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
 import net.minecraftforge.fml.ModLoadingContext;
 import net.minecraftforge.fml.common.Mod;
 import net.minecraftforge.fml.config.ModConfig;
+import net.minecraftforge.fml.event.config.ModConfigEvent;
 import net.minecraftforge.fml.event.lifecycle.FMLConstructModEvent;
+import net.minecraftforge.fml.javafmlmod.FMLJavaModLoadingContext;
 
 import java.util.OptionalInt;
 
@@ -36,58 +38,86 @@ public class ThinAirForge {
     public static void onConstructMod(final FMLConstructModEvent evt) {
         ForgeModRegistry.touch();
         ModLoadingContext.get().registerConfig(ModConfig.Type.COMMON, CommonConfig.SPEC);
+        registerLoadingHandlers(FMLJavaModLoadingContext.get().getModEventBus());
         ModConstructor.construct(ThinAir.MOD_ID, ThinAir::new);
         registerCapabilities();
         // TODO remove when this has been re-enabled in Puzzles Lib
         registerEventInvokers();
         registerIntegrations();
-        DataProviderHelper.registerDataProviders(ThinAir.MOD_ID, ModAdvancementProvider::new, ModBlockLootProvider::new, ModModelProvider::new, ModBlockTagsProvider::new, ModChestLootProvider::new, ModEntityTypeTagsProvider::new, ModRecipeProvider::new);
+        DataProviderHelper.registerDataProviders(ThinAir.MOD_ID,
+                ModAdvancementProvider::new,
+                ModBlockLootProvider::new,
+                ModModelProvider::new,
+                ModBlockTagsProvider::new,
+                ModChestLootProvider::new,
+                ModEntityTypeTagsProvider::new,
+                ModRecipeProvider::new);
         DataProviderHelper.registerDataProviders(ThinAir.MOD_ID, ModItemTagsProvider::new);
+    }
+
+    private static void registerLoadingHandlers(IEventBus eventBus) {
+        eventBus.addListener((final ModConfigEvent evt) -> {
+            if (evt instanceof ModConfigEvent.Loading || evt instanceof ModConfigEvent.Reloading) {
+                if (evt.getConfig().getModId().equals(ThinAir.MOD_ID)
+                        && evt.getConfig().getType() == ModConfig.Type.COMMON) {
+                    CommonConfig.onModConfigLoading();
+                }
+            }
+        });
     }
 
     @Deprecated
     private static void registerEventInvokers() {
-        ForgeEventInvokerRegistry.INSTANCE.register(LivingEvents.Breathe.class, LivingBreatheEvent.class, (LivingEvents.Breathe callback, LivingBreatheEvent evt) -> {
-            final int airAmountValue;
-            if (!evt.canBreathe()) {
-                airAmountValue = -evt.getConsumeAirAmount();
-            } else if (evt.canRefillAir()) {
-                airAmountValue = evt.getRefillAirAmount();
-            } else {
-                airAmountValue = 0;
-            }
-            DefaultedInt airAmount = DefaultedInt.fromValue(airAmountValue);
-            LivingEntity entity = evt.getEntity();
-            // do not use LivingBreatheEvent::canBreathe, it is merged with LivingBreatheEvent::canRefillAir, so recalculate the value
-            boolean canLoseAir = !entity.canDrownInFluidType(entity.getEyeInFluidType()) && !MobEffectUtil.hasWaterBreathing(entity) && (!(entity instanceof Player) || !((Player) entity).getAbilities().invulnerable);
-            EventResult result = callback.onLivingBreathe(entity, airAmount, evt.canRefillAir(), canLoseAir);
-            if (result.isInterrupt()) {
-                evt.setCanBreathe(true);
-                evt.setCanRefillAir(false);
-            } else {
-                OptionalInt optional = airAmount.getAsOptionalInt();
-                if (optional.isPresent()) {
-                    if (optional.getAsInt() < 0) {
-                        evt.setCanBreathe(false);
-                        evt.setConsumeAirAmount(Math.abs(optional.getAsInt()));
+        ForgeEventInvokerRegistry.INSTANCE.register(LivingEvents.Breathe.class,
+                LivingBreatheEvent.class,
+                (LivingEvents.Breathe callback, LivingBreatheEvent evt) -> {
+                    final int airAmountValue;
+                    if (!evt.canBreathe()) {
+                        airAmountValue = -evt.getConsumeAirAmount();
+                    } else if (evt.canRefillAir()) {
+                        airAmountValue = evt.getRefillAirAmount();
                     } else {
-                        evt.setCanBreathe(true);
-                        evt.setCanRefillAir(true);
-                        evt.setRefillAirAmount(optional.getAsInt());
+                        airAmountValue = 0;
                     }
-                }
-            }
-        });
-        ForgeEventInvokerRegistry.INSTANCE.register(LivingEvents.Drown.class, LivingDrownEvent.class, (LivingEvents.Drown callback, LivingDrownEvent evt) -> {
-            EventResult result = callback.onLivingDrown(evt.getEntity(), evt.getEntity().getAirSupply(), evt.isDrowning());
-            if (result.isInterrupt()) {
-                if (result.getAsBoolean()) {
-                    evt.setDrowning(true);
-                } else {
-                    evt.setCanceled(true);
-                }
-            }
-        });
+                    DefaultedInt airAmount = DefaultedInt.fromValue(airAmountValue);
+                    LivingEntity entity = evt.getEntity();
+                    // do not use LivingBreatheEvent::canBreathe, it is merged with LivingBreatheEvent::canRefillAir, so recalculate the value
+                    boolean canLoseAir =
+                            !entity.canDrownInFluidType(entity.getEyeInFluidType()) && !MobEffectUtil.hasWaterBreathing(
+                                    entity) && (!(entity instanceof Player)
+                                    || !((Player) entity).getAbilities().invulnerable);
+                    EventResult result = callback.onLivingBreathe(entity, airAmount, evt.canRefillAir(), canLoseAir);
+                    if (result.isInterrupt()) {
+                        evt.setCanBreathe(true);
+                        evt.setCanRefillAir(false);
+                    } else {
+                        OptionalInt optional = airAmount.getAsOptionalInt();
+                        if (optional.isPresent()) {
+                            if (optional.getAsInt() < 0) {
+                                evt.setCanBreathe(false);
+                                evt.setConsumeAirAmount(Math.abs(optional.getAsInt()));
+                            } else {
+                                evt.setCanBreathe(true);
+                                evt.setCanRefillAir(true);
+                                evt.setRefillAirAmount(optional.getAsInt());
+                            }
+                        }
+                    }
+                });
+        ForgeEventInvokerRegistry.INSTANCE.register(LivingEvents.Drown.class,
+                LivingDrownEvent.class,
+                (LivingEvents.Drown callback, LivingDrownEvent evt) -> {
+                    EventResult result = callback.onLivingDrown(evt.getEntity(),
+                            evt.getEntity().getAirSupply(),
+                            evt.isDrowning());
+                    if (result.isInterrupt()) {
+                        if (result.getAsBoolean()) {
+                            evt.setDrowning(true);
+                        } else {
+                            evt.setCanceled(true);
+                        }
+                    }
+                });
     }
 
     private static void registerCapabilities() {